--- a/binding/form_mapping.go
+++ b/binding/form_mapping.go
@@ -31,7 +31,7 @@ func mapFormByTag(ptr interface{}, form map[string][]string, tag string) error {
 	return mappingByPtr(ptr, formSource(form), tag)
 }
 
-// setter - try to set value on a walking by fields of a struct
+// setter tries to set value on a walking by fields of a struct
 type setter interface {
 	TrySet(value reflect.Value, field reflect.StructField, key string, opt setOptions) (isSetted bool, err error)
 }
@@ -40,7 +40,7 @@ type formSource map[string][]string
 
 var _ setter = formSource(nil)
 
-// TrySet - try to set a value by request's form source (like map[string][]string)
+// TrySet tries to set a value by request's form source (like map[string][]string)
 func (form formSource) TrySet(value reflect.Value, field reflect.StructField, tagValue string, opt setOptions) (isSetted bool, err error) {
 	return setByForm(value, field, form, tagValue, opt)
 }