Refactor mapping to interface `setter` - for supporting a multipart request.

Benchmarks of a simple form mapping (with many fields and one field respectively):
```
⇒  benchcmp old.bch new.bch
benchmark                  old ns/op     new ns/op     delta
BenchmarkMapFormFull-4     2332          2358          +1.11%
BenchmarkMapFormName-4     218           235           +7.80%

benchmark                  old allocs     new allocs     delta
BenchmarkMapFormFull-4     12             12             +0.00%
BenchmarkMapFormName-4     1              1              +0.00%

benchmark                  old bytes     new bytes     delta
BenchmarkMapFormFull-4     200           200           +0.00%
BenchmarkMapFormName-4     8             8             +0.00%
```